+++ otterdog/operations/apply.py
@@ -22,6 +22,7 @@ class ApplyOperation(PlanOperation):
         self,
         force_processing: bool,
         no_web_ui: bool,
+        repo_filter: str,
         update_webhooks: bool,
         update_secrets: bool,
         update_filter: str,
@@ -29,7 +30,7 @@ class ApplyOperation(PlanOperation):
         resolve_secrets: bool = True,
         include_resources_with_secrets: bool = True,
     ):
-        super().__init__(no_web_ui, update_webhooks, update_secrets, update_filter)
+        super().__init__(no_web_ui, repo_filter, update_webhooks, update_secrets, update_filter)
         self._force_processing = force_processing
         self._delete_resources = delete_resources
         self._resolve_secrets = resolve_secrets